# legalai-backend/app/static/openapi.yaml
openapi: 3.0.3
info:
  title: Legal Lawyer AI API
  version: "1.0.0"
  description: |
    API specification for Legal Lawyer AI backend.

    Safe Mode:
    - Only endpoints that declare the `X-Safe-Mode` header enforce it.
    - When `X-Safe-Mode: 1`, those endpoints block writes and return 403.
servers:
  - url: https://gowaned-beckham-unlawful.ngrok-free.dev
  - url: http://127.0.0.1:5000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    safeModeHeader:
      type: apiKey
      in: header
      name: X-Safe-Mode
      description: "Set to '1' to enable safe mode (read-only, no persistence)"
  schemas:
    AnyObject:
      type: object
      additionalProperties: true
    Error:
      type: object
      properties:
        message: { type: string }
        error: { type: string }
        details: { $ref: "#/components/schemas/AnyObject" }
...
    OkIdResponse:
      type: object
      required: [id]
      properties:
        id:
          type: integer
          format: int64
...
    PathwaySchema:
      type: object
      required: [id, title, steps, language]
      properties:
        id: { type: integer, format: int64 }
        title: { type: string, maxLength: 200 }
        summary: { type: string, nullable: true }
        steps:
          type: array
          items:
            type: object
            properties:
              step: { type: integer }
              title: { type: string }
              description: { type: string }
        category: { type: string, maxLength: 120, nullable: true }
        language: { type: string, enum: [en, ur] }
        tags: { type: array, items: { type: string } }
...
    ReminderSchema:
      type: object
      required: [id, title, scheduledAt, timezone, isDone]
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        notes: { type: string, nullable: true }
        scheduledAt: { type: string, format: date-time }
        timezone: { type: string }
        isDone: { type: boolean }

    # Emergency Contact Schemas
...
    UnauthorizedErrorResponse:
      type: object
      required: [message]
      properties:
        message: { type: string, example: "Invalid token" }
        error: { type: string, example: "Unauthorized" }
    ForbiddenErrorResponse:
      type: object
      required: [message]
      properties:
        message: { type: string, example: "Admin only" }
        error: { type: string, example: "Forbidden" }
...
  parameters:
...
    CategoryParam:
      in: query
      name: category
      schema:
        type: string
      description: "Filter by category"
...
security: [] # default none; per-route add as needed
paths:
...
  /api/v1/pathways:
    get:
      tags: [Pathways]
      summary: List pathways
      description: Returns pathways filtered by user's language preference and optional category
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CategoryParam"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/PathwaySchema" }
                  message:
                    type: string
                    nullable: true
              examples:
                withData:
                  summary: Pathways found
                  value:
                    data:
                      - id: 1
                        title: "Divorce Process in Pakistan"
                        summary: "Step-by-step guide"
                        steps:
                          - step: 1
                            title: "File petition"
                            description: "File in family court"
                        category: "family"
                        language: "en"
                        tags: ["divorce"]
                    message: null
                noContent:
                  summary: No content in language
                  value:
                    data: []
                    message: "Content not available in selected language"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
    post:
      tags: [Pathways]
      summary: Create pathway (Admin)
      description: |
        Creates a new legal pathway (step-by-step guide).
        
        **Admin Only**
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, steps]
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
                  example: "Divorce Process in Pakistan"
                summary:
                  type: string
                  nullable: true
                  example: "Step-by-step guide for divorce proceedings"
                steps:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    properties:
                      step: { type: integer, example: 1 }
                      title: { type: string, example: "File petition" }
                      description: { type: string, example: "File in family court" }
                  example:
                    - step: 1
                      title: "File petition"
                      description: "File in family court"
                    - step: 2
                      title: "Notice period"
                      description: "90 days waiting period"
                category:
                  type: string
                  maxLength: 120
                  nullable: true
                  example: "family"
                language:
                  type: string
                  enum: [en, ur]
                  default: en
                tags:
                  type: array
                  items: { type: string }
                  default: []
                  example: ["divorce", "family-law"]
      responses:
        "201":
          description: Pathway created successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkIdResponse" }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/pathways/{pathwayId}:
    get:
      tags: [Pathways]
      summary: Get single pathway
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: pathwayId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [data, message]
                properties:
                  data: { $ref: "#/components/schemas/PathwaySchema" }
                  message: { type: string, nullable: true }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
        "404":
          description: Content not available in selected language
          content:
            application/json:
              schema:
                type: object
                required: [data, message]
                properties:
                  data: { type: object, nullable: true }
                  message: { type: string }
    put:
      tags: [Pathways]
      summary: Update pathway (Admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: pathwayId
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
                summary:
                  type: string
                  nullable: true
                steps:
                  type: array
                  items:
                    type: object
                    properties:
                      step: { type: integer }
                      title: { type: string }
                      description: { type: string }
                category:
                  type: string
                  maxLength: 120
                  nullable: true
                language:
                  type: string
                  enum: [en, ur]
                tags:
                  type: array
                  items: { type: string }
      responses:
        "200":
          description: Pathway updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
        "403":
          description: Forbidden (Admin only)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ForbiddenErrorResponse" }
    delete:
      tags: [Pathways]
      summary: Delete pathway (Admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: pathwayId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }

  # -------------------- CHECKLISTS --------------------
...
  /api/v1/reminders:
    post:
      tags: [Reminders]
      summary: Create reminder
      description: |
        Creates a new reminder with push notification scheduling.
        
        **Safe Mode:** Blocked when `X-Safe-Mode: 1`
      security: [{ bearerAuth: [] }, { safeModeHeader: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, scheduledAt]
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
                  example: "Court Hearing"
                notes:
                  type: string
                  nullable: true
                  example: "Family court hearing at 10 AM"
                scheduledAt:
                  type: string
                  format: date-time
                  example: "2025-01-20T10:00:00Z"
                timezone:
                  type: string
                  example: "Asia/Karachi"
      responses:
        "201":
          description: Reminder created successfully
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkIdResponse" }
        "400":
          description: Validation error (missing title or scheduledAt)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
        "403":
          description: Safe mode enabled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    get:
      tags: [Reminders]
      summary: List reminders
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ReminderSchema" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
  /api/v1/reminders/{reminderId}:
    put:
      tags: [Reminders]
      summary: Update reminder
      description: |
        Updates reminder details or marks as done.
        
        **Safe Mode:** Blocked when `X-Safe-Mode: 1`
      security: [{ bearerAuth: [] }, { safeModeHeader: [] }]
      parameters:
        - in: path
          name: reminderId
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
                notes:
                  type: string
                  nullable: true
                scheduledAt:
                  type: string
                  format: date-time
                timezone:
                  type: string
                isDone:
                  type: boolean
            example:
              title: "Updated Court Hearing"
              isDone: true
      responses:
        "200":
          description: Reminder updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
        "400":
          description: Validation error
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
        "403":
          description: Forbidden (Safe mode or not owner)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: Reminder not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      tags: [Reminders]
      summary: Delete reminder
      security: [{ bearerAuth: [] }, { safeModeHeader: [] }]
      parameters:
        - in: path
          name: reminderId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
        "403":
          description: Safe mode enabled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/reminders/register-device-token:
    post:
      tags: [Reminders]
      summary: Register device token
      description: |
        Registers FCM/APNS token for push notifications.
        
        Upserts token (deletes old, creates new).
        
        **Safe Mode:** Blocked when `X-Safe-Mode: 1`
      security: [{ bearerAuth: [] }, { safeModeHeader: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform, token]
              properties:
                platform:
                  type: string
                  enum: [android, ios]
                  example: "android"
                token:
                  type: string
                  minLength: 10
                  example: "fcm_device_token_here_very_long_string..."
      responses:
        "200":
          description: Token registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
        "400":
          description: Invalid platform or missing token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
        "403":
          description: Safe mode enabled
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # -------------------- NOTIFICATIONS --------------------
...